backend/requirements/base.txt
Django==2.0.4
psycopg2-binary==2.7.4
django-crispy-forms==1.7.2
django-webpack-loader==0.6.0
django-compressor==2.2
django-gravatar2==1.4.2
djangorestframework==3.8.2
#django-allauth==0.37.1
django-smuggler==0.9.0
dropbox==9.0.0
